+++ 1DGratingAnalysis.py
@@ -3,64 +3,7 @@ import numpy as np
 import scipy.signal as sig
 import Functions.Maths as math
 import Functions.Organisation as org
-from PIL import Image
-
-
-def readSemLog(file_path):
-    '''
-    Read .txt file output from SEM image output. Removes $ and \n characters
-    from line.
-    Args:
-        file_path: <string> path to file
-    Returns:
-        line: <array> array of stripped lines
-    '''
-    with open(file_path) as infile:
-        alllines = infile.readlines()
-        lines = [
-            (line.translate({ord('$'): None}))[0: -1]
-            for line in alllines]
-    return lines
-
-
-def SEMDict(lines):
-    '''
-    Read lines from SEM output txt file and pulls important parameters into
-    dictionary. To further add parameters, simply add names and keys to the
-    sample dictionary in this function. Remember FULL_SIZE parameter is [width,
-    height].
-    Args:
-        lines: <array> array of lines from txt file, stripped of '$' and '\n'
-    Returns:
-        parameterdict: <dict> parameter dictionary
-    '''
-    parameterdict = {
-        'SM_EMI_CURRENT': None,
-        'CM_ACCEL_VOLT': None,
-        'SM_WD': None,
-        'CM_MAG': None,
-        'CM_BRIGHTNESS': None,
-        'CM_CONTRAST': None,
-        'CM_FULL_SIZE': None,
-        'SM_MICRON_BAR': None,
-        'SM_MICRON_MARKER': None}
-    for line in lines:
-        splitline = line.split(' ')
-        if splitline[0] in parameterdict.keys():
-            parameterdict[f'{splitline[0]}'] = splitline[1:]
-    return parameterdict
-
-
-def OpenImgFile(file_path):
-    '''
-    Loads image file.
-    Args:
-        file_path: <string> path to file
-    Returns:
-        image: <object> image file for PIL
-    '''
-    image = Image.open(file_path)
-    return image
+import src.fileIO as fileIO
 
 
 def ImageRegionInterest(image,
@@ -186,13 +129,15 @@ if __name__ == '__main__':
             txtpath = os.path.join(
                 dirpath,
                 file)
-            lines = readSemLog(file_path=txtpath)
-            semParameters = SEMDict(lines=lines)
+
+            lines = fileIO.readSemLog(file_path=txtpath)
+            semParameters = fileIO.SEMDict(lines=lines)
 
             imagepath = os.path.join(
                 dirpath,
                 f'{file[0: -4]}.bmp')
-            image = OpenImgFile(file_path=imagepath)
+
+            image = fileIO.OpenImgFile(file_path=imagepath)
 
             pixels, roi = ImageRegionInterest(
                 image=image,
@@ -228,7 +173,7 @@ if __name__ == '__main__':
                         math.StandardErrorMean(x=f)
                         for f in np.array(frequencies).T]})
 
-            org.saveJson(
+            fileIO.saveJson(
                 out_path=os.path.join(
                     dirpath,
                     f'{file[0: -4]}_Results.csv'),
